size t not int

+++ 14-binary_tree_balance.c
@@ -4,7 +4,7 @@
  * @tree: pointer to root node
  * Return: measure of balance or 0 if NULL
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
 size_t left_h = 0, right_h = 0;
 if (tree == NULL)
